refactor(utils): Move validation utilities to validation/ subdirectory
Phase 2 (Day 1): Validation utilities
- Moved 4 validation files to src/utils/validation/
- Updated 2 source imports + 2 test imports
- Fixed csv-processor-core.ts relative path
- Tests passing (199/199 smoke tests)

Files moved:
- isbn-validation.ts
- book-validation.ts
- json-validator.ts
- csv-validator.ts
